# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/config/plan": {
      "post": {
        "description": "Diffs the desired config against the current rules and returns the plan (operations, summary, errors, warnings) without mutating anything. A missing `desired` key returns 400; invalid desired rules surface as plan errors (still 200). Returns 200 (NestJS would otherwise default POST to 201).",
        "operationId": "ConfigPlanController_plan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPlanBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The config plan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigPlanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "`desired` is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Plan a config change",
        "tags": [
          "config"
        ]
      }
    },
    "/api/config/import": {
      "post": {
        "description": "Imports a Portier config in `replace` (replace all rules) or `merge` (add non-conflicting rules) mode. Returns 200 with the import result + the full rule list on success, or 422 with the errors + result when the import reports validation/duplicate-binding/merge-conflict errors (rejected before any mutation). An invalid mode or config returns 400. Unlike `plan` (a dry-run), import MUTATES the rules; enabled imported rules start their forwarders.",
        "operationId": "ConfigImportController_import",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigImportBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import succeeded; the full rule list is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigImportResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid mode or config.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "The import reported errors (no mutation).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigImportErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Import a config",
        "tags": [
          "config"
        ]
      }
    },
    "/api/config/apply": {
      "post": {
        "description": "Diffs the desired config against the current rules and applies it. A missing `desired` key returns 400; a plan with errors returns 200 `ok:false` (no mutation). `dryRun:true` returns the plan + applied counts without mutating (and is allowed even for destructive plans). A non-dry-run plan with destructive operations requires `yes:true` (else 400). When the plan has drift, the rules are replaced (MUTATING; enabled rules start). Unlike `plan` (always a dry-run) and `import` (a raw replace/merge), apply gates on drift/destructive/confirmation and returns the plan with an `appliedAt` timestamp.",
        "operationId": "ConfigApplyController_apply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigApplyBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The apply result (ok:true applied / ok:true dry-run / ok:false on plan errors), with the plan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigApplyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "`desired` is required, or destructive operations require `yes:true`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Apply a config change",
        "tags": [
          "config"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portier API",
    "description": "Portier management API — generated from the NestJS controllers and DTOs that serve the `/api` surface.",
    "version": "1.14",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "errors": {
            "description": "One or more human-readable error messages.",
            "example": [
              "port must be an integer from 1 to 65535."
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "errors"
        ]
      },
      "ConfigAppliedCountsDto": {
        "type": "object",
        "properties": {
          "add": {
            "type": "number",
            "description": "Rules added."
          },
          "update": {
            "type": "number",
            "description": "Rules updated."
          },
          "remove": {
            "type": "number",
            "description": "Rules removed."
          },
          "unchanged": {
            "type": "number",
            "description": "Rules unchanged."
          }
        },
        "required": [
          "add",
          "update",
          "remove",
          "unchanged"
        ]
      },
      "ConfigApplyBodyDto": {
        "type": "object",
        "properties": {
          "desired": {
            "description": "The desired config (object with a rules array). Required.",
            "allOf": [
              {
                "$ref": "#/components/schemas/DesiredConfigDto"
              }
            ]
          },
          "yes": {
            "type": "boolean",
            "description": "Confirm destructive operations. Required (true) when the plan is destructive."
          },
          "dryRun": {
            "type": "boolean",
            "description": "Preview only — return the plan + applied counts without mutating."
          }
        },
        "required": [
          "desired"
        ]
      },
      "ConfigApplyResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Whether the apply succeeded (false when the plan has errors)."
          },
          "dryRun": {
            "type": "boolean",
            "description": "Whether this was a dry-run (no mutation)."
          },
          "appliedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the apply ran (ISO 8601)."
          },
          "plan": {
            "description": "The plan that was applied (or would be).",
            "allOf": [
              {
                "$ref": "#/components/schemas/ConfigPlanResponseDto"
              }
            ]
          },
          "applied": {
            "description": "The applied operation counts.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ConfigAppliedCountsDto"
              }
            ]
          }
        },
        "required": [
          "ok",
          "dryRun",
          "appliedAt",
          "plan",
          "applied"
        ]
      },
      "ConfigExportResponseDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "enum": [
              "1"
            ],
            "description": "Export schema version."
          },
          "exportedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the config was exported."
          },
          "rules": {
            "description": "Exported forward rules.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ForwardRuleDto"
            }
          }
        },
        "required": [
          "version",
          "exportedAt",
          "rules"
        ]
      },
      "ConfigImportBodyDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "replace",
              "merge"
            ],
            "description": "Import mode: replace all rules, or merge non-conflicting rules."
          },
          "config": {
            "description": "The Portier config to import (version 1 + a rules array).",
            "allOf": [
              {
                "$ref": "#/components/schemas/ConfigExportResponseDto"
              }
            ]
          }
        },
        "required": [
          "mode",
          "config"
        ]
      },
      "ConfigImportErrorResponseDto": {
        "type": "object",
        "properties": {
          "errors": {
            "description": "Import error messages.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "result": {
            "description": "The import outcome (zero imported on error).",
            "allOf": [
              {
                "$ref": "#/components/schemas/ImportResultDto"
              }
            ]
          }
        },
        "required": [
          "errors",
          "result"
        ]
      },
      "ConfigImportResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "description": "The import outcome counts.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ImportResultDto"
              }
            ]
          },
          "rules": {
            "description": "The full rule list after import (with advisories).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ForwardRuleResponseDto"
            }
          }
        },
        "required": [
          "result",
          "rules"
        ]
      },
      "ConfigPlanBodyDto": {
        "type": "object",
        "properties": {
          "desired": {
            "description": "The desired config (object with a rules array). Required.",
            "allOf": [
              {
                "$ref": "#/components/schemas/DesiredConfigDto"
              }
            ]
          }
        },
        "required": [
          "desired"
        ]
      },
      "ConfigPlanChangeDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "The changed material field."
          },
          "before": {
            "type": "object",
            "additionalProperties": true,
            "description": "Value before the change (a JSON scalar)."
          },
          "after": {
            "type": "object",
            "additionalProperties": true,
            "description": "Value after the change (a JSON scalar)."
          }
        },
        "required": [
          "field",
          "before",
          "after"
        ]
      },
      "ConfigPlanErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "field": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ConfigPlanOperationDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "add",
              "update",
              "remove",
              "unchanged"
            ],
            "description": "Operation type."
          },
          "ruleId": {
            "type": "string",
            "description": "Affected rule id, if known."
          },
          "ruleName": {
            "type": "string"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "tcp",
              "udp"
            ]
          },
          "current": {
            "description": "Current rule snapshot.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ConfigPlanRuleSnapshotDto"
              }
            ]
          },
          "desired": {
            "description": "Desired rule snapshot.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ConfigPlanRuleSnapshotDto"
              }
            ]
          },
          "changes": {
            "description": "Field-level changes (update only).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigPlanChangeDto"
            }
          },
          "destructive": {
            "type": "boolean",
            "description": "Whether the operation restarts/removes a forwarder."
          }
        },
        "required": [
          "type",
          "ruleName",
          "protocol",
          "destructive"
        ]
      },
      "ConfigPlanResponseDto": {
        "type": "object",
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the plan was generated (ISO 8601)."
          },
          "mode": {
            "type": "string",
            "enum": [
              "plan"
            ],
            "description": "Discriminator."
          },
          "summary": {
            "$ref": "#/components/schemas/ConfigPlanSummaryDto"
          },
          "operations": {
            "description": "Per-rule operations.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigPlanOperationDto"
            }
          },
          "errors": {
            "description": "Plan validation errors.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigPlanErrorDto"
            }
          },
          "warnings": {
            "description": "Plan advisory warnings.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigPlanWarningDto"
            }
          }
        },
        "required": [
          "generatedAt",
          "mode",
          "summary",
          "operations",
          "errors",
          "warnings"
        ]
      },
      "ConfigPlanRuleSnapshotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Rule id (absent for an added rule)."
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "tcp",
              "udp"
            ]
          },
          "listenHost": {
            "type": "string"
          },
          "listenPort": {
            "type": "number"
          },
          "targetHost": {
            "type": "string"
          },
          "targetPort": {
            "type": "number"
          },
          "enabled": {
            "type": "boolean"
          },
          "udpMode": {
            "type": "string",
            "enum": [
              "one-way",
              "bidirectional-last-client",
              "bidirectional-multi-client"
            ]
          },
          "group": {
            "type": "string",
            "description": "Optional group label."
          }
        },
        "required": [
          "name",
          "protocol",
          "listenHost",
          "listenPort",
          "targetHost",
          "targetPort",
          "enabled"
        ]
      },
      "ConfigPlanSummaryDto": {
        "type": "object",
        "properties": {
          "add": {
            "type": "number"
          },
          "update": {
            "type": "number"
          },
          "remove": {
            "type": "number"
          },
          "unchanged": {
            "type": "number"
          },
          "destructive": {
            "type": "number",
            "description": "Number of destructive operations."
          },
          "hasDrift": {
            "type": "boolean",
            "description": "Whether the desired config differs from current."
          },
          "hasErrors": {
            "type": "boolean",
            "description": "Whether the plan has validation errors."
          }
        },
        "required": [
          "add",
          "update",
          "remove",
          "unchanged",
          "destructive",
          "hasDrift",
          "hasErrors"
        ]
      },
      "ConfigPlanWarningDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "DesiredConfigDto": {
        "type": "object",
        "properties": {
          "rules": {
            "description": "Desired rule snapshots.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigPlanRuleSnapshotDto"
            }
          }
        },
        "required": [
          "rules"
        ]
      },
...
      "ForwardRuleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable rule id."
          },
          "name": {
            "type": "string",
            "description": "Operator-facing rule name."
          },
          "protocol": {
            "type": "string",
            "enum": [
              "tcp",
              "udp"
            ],
            "description": "Forwarding protocol."
          },
          "listenHost": {
            "type": "string",
            "description": "Local listen host.",
            "example": "127.0.0.1"
          },
          "listenPort": {
            "type": "number",
            "description": "Local listen port.",
            "example": 48010
          },
          "targetHost": {
            "type": "string",
            "description": "Forward target host."
          },
          "targetPort": {
            "type": "number",
            "description": "Forward target port."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the rule autostarts."
          },
          "udpMode": {
            "type": "string",
            "enum": [
              "one-way",
              "bidirectional-last-client",
              "bidirectional-multi-client"
            ],
            "description": "UDP forwarding mode (UDP rules only)."
          },
          "group": {
            "type": "string",
            "description": "Optional grouping label."
          }
        },
        "required": [
          "id",
          "name",
          "protocol",
          "listenHost",
          "listenPort",
          "targetHost",
          "targetPort",
          "enabled"
        ]
      },
      "ForwardRuleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable rule id."
          },
          "name": {
            "type": "string",
            "description": "Operator-facing rule name."
          },
          "protocol": {
            "type": "string",
            "enum": [
              "tcp",
              "udp"
            ],
            "description": "Forwarding protocol."
          },
          "listenHost": {
            "type": "string",
            "description": "Local listen host.",
            "example": "127.0.0.1"
          },
          "listenPort": {
            "type": "number",
            "description": "Local listen port.",
            "example": 48010
          },
          "targetHost": {
            "type": "string",
            "description": "Forward target host."
          },
          "targetPort": {
            "type": "number",
            "description": "Forward target port."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the rule autostarts."
          },
          "udpMode": {
            "type": "string",
            "enum": [
              "one-way",
              "bidirectional-last-client",
              "bidirectional-multi-client"
            ],
            "description": "UDP forwarding mode (UDP rules only)."
          },
          "group": {
            "type": "string",
            "description": "Optional grouping label."
          },
          "advisories": {
            "description": "Port advisories for the rule's listen binding.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortAdvisoryDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "protocol",
          "listenHost",
          "listenPort",
          "targetHost",
          "targetPort",
          "enabled",
          "advisories"
        ]
      },
...
      "ImportResultDto": {
        "type": "object",
        "properties": {
          "imported": {
            "type": "number",
            "description": "Number of rules imported."
          },
          "skipped": {
            "type": "number",
            "description": "Number of rules skipped (merge conflicts)."
          },
          "errors": {
            "description": "Import error messages (empty on success).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "imported",
          "skipped",
          "errors"
        ]
      },
      "PortAdvisoryDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "COMMON_PORT",
              "PRIVILEGED_PORT",
              "OUTSIDE_RECOMMENDED_RANGE",
              "LAN_EXPOSURE",
              "MANAGEMENT_LAN_EXPOSURE"
            ],
            "description": "Advisory code."
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "danger"
            ],
            "description": "Advisory severity."
          },
          "message": {
            "type": "string",
            "description": "Human-readable advisory message."
          }
        },
        "required": [
          "code",
          "severity",
          "message"
        ]
      },
...
    }
  }
}